requirements.txt
django>=5.1.4
djangorestframework>=3.15.2
django-cors-headers>=4.6.0
mysqlclient>=2.2.7
python-dotenv>=1.0.0
Pillow>=10.2.0
requests>=2.31.0
pandas>=2.2.3
konlpy>=0.6.0
scikit-learn>=1.6.0
numpy>=2.2.1
scipy>=1.15.0
JPype1>=1.5.1
lxml>=5.3.0
joblib>=1.4.2
threadpoolctl>=3.5.0
python-dateutil>=2.9.0
pytz>=2024.2
tzdata>=2024.2
